requirements.txt
# 核心依赖
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.0.0
PyYAML>=6.0

# LLM和API相关
openai>=1.0.0
requests>=2.28.0
transformers>=4.20.0
torch>=1.12.0

# Web界面
gradio>=3.50.0

# 数据处理
scikit-learn>=1.1.0
nltk>=3.8
spacy>=3.4.0

# 评估和指标
rouge-score>=0.1.2
bert-score>=0.3.11

# 可视化和报告
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# 工具和实用程序
tqdm>=4.64.0
python-dotenv>=0.19.0
click>=8.1.0

# 开发和测试
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# 可选：本地模型支持
# sentence-transformers>=2.2.0
# accelerate>=0.20.0
# bitsandbytes>=0.39.0 